Split Handler::Exifr into Handler::ExifrJPEG and Handler::Exifr::TIFF
