+++ lib/xmp.rb
@@ -3,7 +3,8 @@ module XMP
   require 'xmp/document'
   require 'xmp/error'
   require 'xmp/handler'
-  require 'xmp/handler/exifr'
+  require 'xmp/handler/exifr_jpeg'
+  require 'xmp/handler/exifr_tiff'
   require 'xmp/handler/file'
   require 'xmp/handler/xml'
   require 'xmp/namespace'
@@ -14,7 +15,8 @@ module XMP
     Handler::File.new('.jpg', '.jpeg') { |file| EXIFR::JPEG.new(file) },
     Handler::File.new('.tif', '.tiff') { |file| EXIFR::TIFF.new(file) },
     Handler::File.new('.xmp', '.xml')  { |file| Nokogiri::XML(file) },
-    Handler::Exifr.new,
+    Handler::ExifrJPEG.new,
+    Handler::ExifrTIFF.new,
     Handler::XML.new
   ]
 end
